Bump mock from 2.0.0 to 3.0.5
grep "from mock" and grep "import mock" and grep Mock in total found two
occurances where Mock is used:
  askbot/deps/group_messaging/tests.py:
  - from mock import Mock
  - request = Mock(spec=spec)
  askbot/tests/user_views_tests.py:
  - from mock import Mock
  - request = Mock(spec=('path', 'REQUEST', 'user'))

This use of Mock is very much in line with the simplest use cases of Mock() as
described in its documentation. It does not make sense to assume there could be
an impediment. If there is, askbot's own group_messaging and user_views tests
will fail. I accept this as sufficient tests and will simply move to
Mock 3.0.5.