Bump unidecode from 0.4.19 to 1.0.23
The main change is the change in numbering versions. The real increase is indicated by the micro version.
19->20->21->22->23. We are skipping 3 releases which introduced additional characters that unidecode can now handle
properly. Everything seems unproblematic, however this function does see user input. The 0.4.19 has 42 tests which
version 1.0.23 all passes. The 1.0.23 has 45 tests of which 0.4.19 passes the first 42. The 3
failures test characters for which support was only added after 0.4.19. Seems to be a straight forward upgrade.