Bump python-cas from 1.1.0 to 1.4.0
Upgrading this is probably a prerequisite for upgrading django, as cas is some authentication library that used to be
django internal and is now a seperate project. Luckily we have only one occurance of this lib in:
askbot/deps/django_authopenid/providers/cas_provider.py

Tracing changes in this project is a bit more difficult than with the previous projects. It would seem there are 31
commits between the currently used version and the most recent release (and another 34 commits unreleased!)
Fun Fact: evgenyfadeev seems involved with this project.

Essentially the 31 released commits are bug fixes and code quality improvements. The interface seems to be unchanged.
Bumping version, moving on.